pyproject.toml
[project]
name = "ask-galadirel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastembed>=0.7.3",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.32",
    "langchain-qdrant>=0.2.0",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
]
